Move NF Simulation and Promoter state analysis to Repository
First attempt at storing previously made code to a github repository. These files use SSA to simulate negative feedback gene circuits. The NF_sim_Promoter_state_analysis_w_plotting file runs the simulations and then stores the amounts of time the promoter is in the on state for a chosen window of time and does this across the length of the simulation. It then creates a dataframe that stores the total time the promoter is on divided by the total window of time and saves it. The fitting files are then used to load these saved dataframes and generate a beta distribution based on the values in the array. It then plots the generated beta distribution against the data. It also performs a Kolmogorov-Smirnov test to compare the data from the simulations to the generated beta distribution.